requirements.txt
# Core framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database and async support
asyncpg>=0.29.0
psycopg2-binary>=2.9.5

# AI and LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langsmith>=0.0.69

# OpenAI and embeddings
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0

# HTTP and API clients
requests>=2.31.0
httpx>=0.25.0

# Monitoring and logging
prometheus-client>=0.17.0
psutil>=5.9.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data processing
python-dateutil>=2.8.0
pytz>=2023.3

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Visualization and plotting
plotly>=5.17.0
matplotlib>=3.7.0

# System utilities
schedule>=1.2.0
